minimal-requirements.txt
docstring==0.1.2.4
PyYAML==6.0.2
Requests==2.32.3
urllib3==2.3.0
